requirements.txt
pypdf
sentence-transformers
faiss-cpu
groq
python-dotenv
pillow
pytesseract